Refactor Azure transfer
